pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Bryan Guan <bryguan@berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {version = "^0.112.2", extras = ["standard"]}
joblib = "1.4.2"
fastapi-cache2 = {version = "0.1.9", extras = ["redis"]}
scikit-learn = "1.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
